Fixed a couple of warnings, updated qt version to 11.2 cause of bug with visual studio

+++ src/baseapp/mvpmodel.cpp
@@ -66,7 +66,7 @@ void MvpModel::export_hair_to_disk(const QString& filename, const HairData& hair
 		throw std::runtime_error("The selected growthmesh could not be found.");
 
 	const int num_hair = hairdata.m_num_hair;
-	const int num_hair_vertices = hairdata.m_vertices.size();
+	const size_t num_hair_vertices = hairdata.m_vertices.size();
 
 	if (num_hair != m_fbx_model_->m_meshes.at(growthmesh_index).m_vertices.size())
 		throw std::runtime_error("The number of hair doesn't match the number of mesh vertices of the growthmesh.");
@@ -74,15 +74,15 @@ void MvpModel::export_hair_to_disk(const QString& filename, const HairData& hair
 	const int num_faces = m_fbx_model_->m_meshes.at(growthmesh_index).get_num_faces();
 
 	std::vector<unsigned int> face_indices = m_fbx_model_->m_meshes.at(growthmesh_index).get_indices();
-	const int num_indices = face_indices.size();
+	const size_t num_indices = face_indices.size();
 
 	std::vector<glm::vec2> uvs = m_fbx_model_->m_meshes.at(growthmesh_index).get_face_uvs();
-	const int num_uvs = uvs.size();
+	const size_t num_uvs = uvs.size();
 
 	if (num_uvs != num_indices)
 		throw std::runtime_error("The number of uv's doesn't match the number of indices.");
 
-	const int num_bones = m_fbx_model_->get_num_bones() == 0 ? 1 : m_fbx_model_->get_num_bones();
+	const size_t num_bones = m_fbx_model_->get_num_bones() == 0 ? 1 : m_fbx_model_->get_num_bones();
 	std::vector<int> bone_indices;
 	std::vector<float> bone_weights;
